A hacky way to create a way to filter categories based on name
I couldn't find a way to filter based on a foreign field within a query that was populated once called. I made the API simply take the ID of the registered article's category and made a new field called "categoryname" that gets filled automatically through the API with it's associated categoryname.